Introduce required changes to make the app work
It is a little bit involved, but this proves a point I'll make sure to
make in class: using Spring is a nightmare. That's why as a developer
I've hated it for so long.

When putting together all the moving parts involved we need to:
- Make sure we don't miss any dependency, required either at compile
  time or runtime.
- Make sure that the versions of the dependencies are compatible with
  each other.
- Make sure that the dependencies are also byte-code compatible with our
  runtime. We might find that the compiler version we are using is
  generating bytecode that cannot be put together at runtime with that
  provided by the dependencies (as happens in this project).
- Make sure that we nail the configuration in every sense. That involves
  several xml files and quite a few annotations to say the least if we
  want this to work on Tomcat.

Needless to say that you need a lot of time and a good amount of
experience to do everything right. Something not easy specially when
working under pressure and tight deadlines.

And this is one of the reasons why Spring Boot was made: to make Spring
usable. But that is another story that we will cover in depth at the end
of Module 5.

For now, let's learn the lesson: Spring is powerful, but complex.